minor changes regarding threaded output

+++ src/include/matrix_vector_kernel.hpp
@@ -10,8 +10,8 @@
 #include <vector>
 #include <fp/fp_blas.hpp>
 
-using real_t = float;
-//using real_t = double;
+//using real_t = float;
+using real_t = double;
 
 #if defined(_BE)
 static constexpr std::int32_t BE = _BE;
